fix: move interop test to its own plan
