--- plans/_integrations/shim/shim.go
@@ -1,5 +1,0 @@
-package shim
-
-func Version() string {
-	return "not-set"
-}
