+++ plans/_integrations_interop/manifest.toml
@@ -1,0 +1,18 @@
+name = "integrations"
+
+[defaults]
+builder = "docker:go"
+runner = "local:docker"
+
+[builders."docker:go"]
+enabled = true
+
+[runners."local:docker"]
+enabled = true
+
+[[testcases]]
+name = "issue-1337-override-builder-configuration"
+instances = { min = 2, max = 2, default = 2 }
+
+[testcases.params]
+  expected_version = { type = "string", desc = "expected version" }
